add: moved Forms to matching pages
