+++ src/app/(main)/create/page.tsx
@@ -1,4 +1,4 @@
-import CreateMeetForm from '@/components/meet-form/CreateMeetForm';
+import CreateMeetForm from '@/app/(main)/create/CreateMeetForm';
 
 export default function CreatePage() {
   return <CreateMeetForm />;
