+++ src/app/(main)/edit/[meetId]/EditMeetForm.tsx
@@ -9,7 +9,7 @@ import {
   MeetTimeInput,
   ConfirmTimeInput,
   MeetPasswordInput,
-} from './inputs';
+} from '@/components/meet-form/inputs';
 
 async function clientAction(formData: FormData) {
   const meetName = formData.get('meetName')?.toString();
